refactor: refactor struct internal
