+++ cmd/api/routes.go
@@ -2,10 +2,10 @@ package main
 
 import (
 	docs "github.com/Ze-Victor/search-zip-code/docs"
-	metrics "github.com/Ze-Victor/search-zip-code/internal/metrics"
 	auth "github.com/Ze-Victor/search-zip-code/internal/pkg/authorization"
 	cep "github.com/Ze-Victor/search-zip-code/internal/pkg/cep"
 	health "github.com/Ze-Victor/search-zip-code/internal/pkg/health"
+	metrics "github.com/Ze-Victor/search-zip-code/internal/pkg/metrics"
 	"github.com/gin-gonic/gin"
 	swaggerfiles "github.com/swaggo/files"
 	ginSwagger "github.com/swaggo/gin-swagger"
